requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
pydantic
slowapi
google-generativeai
PyPDF2
python-docx
python-multipart
python-dotenv
alembic
